cosmetic: removed unneeded cast
git-svn-id: https://xbmc.svn.sourceforge.net/svnroot/xbmc/branches/linuxport/XBMC@20714 568bbfeb-2a22-0410-94d2-cc84cf5bfa90

+++ guilib/GUITextBox.cpp
@@ -319,7 +319,7 @@ unsigned int CGUITextBox::GetRows() const
 
 int CGUITextBox::GetCurrentPage() const
 {
-  if (m_offset + m_itemsPerPage >= (int)GetRows())  // last page
+  if (m_offset + m_itemsPerPage >= GetRows())  // last page
     return (GetRows() + m_itemsPerPage - 1) / m_itemsPerPage;
   return m_offset / m_itemsPerPage + 1;
 }
